AI is done! But bugged
AI now works.
need to add halt to randomizer (all of the AI's receive the same
output). consider swapping to time based randomization, or ensuring last
random output is not like the one before.
Also, AI's can eat each other, in which case when 2 are eaten an
infinite (and unknown) loop is introduced and game gets stuck. easiest
solution is checking for ghosts collision.